requirements.txt
addict==2.4.0
appdirs==1.4.4
audioread==3.0.1
av==12.0.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
docker-pycreds==0.4.0
einops==0.7.0
filelock==3.13.3
fsspec==2024.3.1
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.22.1
Hydra==2.5
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
iopath==0.1.10
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
mmcv-full==1.7.0
moviepy==1.0.3
msgpack==1.0.8
networkx==3.1
numba==0.58.1
numpy==1.24.4
opencv-python==4.9.0.80
packaging==24.0
parameterized==0.9.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
portalocker==2.8.2
proglog==0.1.10
protobuf==4.25.3
psutil==5.9.8
pycparser==2.21
pytorchvideo==0.1.5
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==1.44.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soxr==0.3.7
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.4.0
timm==0.9.16
tomli==2.0.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
wandb==0.16.6
wget==3.2
yacs==0.1.8
yapf==0.40.2
zipp==3.18.1
